+++ src/lib/svg/languages/renderers/default.ts
@@ -1,4 +1,5 @@
-import { ThemeConfig, LanguageStat, LangChartOptions } from "@/lib/types";
+import { LanguageStat, LangChartOptions } from "@/lib/types";
+import { ThemeConfig } from "@/lib/themes/types";
 import { escapeXml } from "@/lib/sanitize";
 import { truncateToWidth, formatLangPct } from "../../text-metrics";
