+++ src/lib/svg/languages/renderers/donut.ts
@@ -1,4 +1,6 @@
-import { ThemeConfig, LanguageStat, LangChartOptions } from "@/lib/types";
+import { LanguageStat, LangChartOptions } from "@/lib/types";
+import { ThemeConfig } from "@/lib/themes/types";
+
 import { escapeXml } from "@/lib/sanitize";
 
 export default function renderDonutLanguageChart(
